+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `--fullpath` flag added to `maelstrom-go-test`
 - `extra-test-binary-args` option added to `maelstrom-go-test`
 - `extra-test-binary-args` option added to `cargo-maelstrom`
+- `extra-pytest-args` option added to `maelstrom-pytest`
 
 #### Changed
 - `ui` test-runner configuration option value now matches the same casing as the CLI option